Added preliminary Clojure 1.3 support to the datastore implementation.
Clojure 1.3 changed the format of metadata tags. ^:key no longer generates metadata of the form {:tag :key}, but {:key true}. This allows multiple tags to exist on the same data item, but breaks simple checks for which tags are defined on a data item.

For now, this patch just checks the running version of Clojure and substitutes the appropriate check.